adding switch on/off for reactions.

+++ src/bns_nurates.h
@@ -261,10 +261,10 @@ struct M1Quantities {
 typedef struct M1Quantities M1Quantities;
 
 struct OpacityFlags {
-  double use_abs_em;
-  double use_pair;
-  double use_brem;
-  double use_iso;
+  int use_abs_em;
+  int use_pair;
+  int use_brem;
+  int use_iso;
 };
 typedef struct OpacityFlags OpacityFlags;
 static OpacityFlags opacity_flags_default_all = {.use_abs_em = 1., .use_pair = 1., .use_brem = 1., .use_iso = 1.};
